wce[2023-08-27]:

+++ wce/2023-08-26/wce.sh
@@ -10,6 +10,7 @@
 WCE_FILE="$HOME/.wce"
 WCE_COUNT_FILE="$HOME/.wce_count"
 WCE_SUSPEND_FILE="$HOME/.wce_suspend"
+WCE_WORKING_DIR="$(cd "$(readlink -f "$(dirname "$0")")/../" || exit 1; pwd)"
 TODAY="$(date +%Y-%m-%d)"
 PS1_ORIGINAL="${PS1}"
 
@@ -19,18 +20,27 @@ function help() {
 }
 
 function start() {
-    if [ -f "${WCE_FILE}" ]; then
+    if [[ -f "${WCE_FILE}" ]]; then
         echo "Error: session already started"
         exit 1
     fi
-    touch "${WCE_FILE}"
-    if [ ! -f "${WCE_COUNT_FILE}" ]; then
+    echo "${TODAY}" > "${WCE_FILE}"
+    if [[ ! -f "${WCE_COUNT_FILE}" ]]; then
         echo "0" > "${WCE_COUNT_FILE}"
         days=0
-    else
-      days="$(cat "${WCE_COUNT_FILE}")"
     fi
+    days=$(($(cat "${WCE_COUNT_FILE}") + 1))
+
+    cd "${WCE_WORKING_DIR}" || echo "Error: can't cd to ${WCE_WORKING_DIR}" && exit 1
+    if [[ ! -d "${TODAY}" ]]; then
+        mkdir "${TODAY}"
+    fi
+    if ! [[ -f "${TODAY}/README.md" ]]; then
+      echo "# ${TODAY}" > "${TODAY}/README.md"
+    fi
+
     echo "Today is ${TODAY}"
+    echo days > "${WCE_COUNT_FILE}"
     echo "Write Code Everyday: The day ${days}"
     echo "Session started!"
     export PS1="wce[${TODAY}]: ${PS1_ORIGINAL}"
@@ -41,15 +51,37 @@ function finish() {
       echo "Error: session not started"
       exit 1
   fi
+  if [[ -f "${WCE_SUSPEND_FILE}" ]]; then
+      echo "Error: session suspended"
+      exit 1
+  fi
+  if [[ $(cat "${WCE_FILE}") != "${TODAY}" ]]; then
+      echo "Error: session not started today"
+      echo "First run 'wce.sh cancel' to cancel the session"
+      exit 1
+  fi
   git add .
   read -r -p "Commit message: " message
   git commit -a -m "wce[${TODAY}]: ${message}"
   # ask wanna push?
   read -r -p "Push to GitHub? [y/N] " response
-  if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
-      git push
+  if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]; then git push
   fi
   echo "Session finished"
+  rm "${WCE_FILE}"
+  export PS1="${PS1_ORIGINAL}"
+}
+
+function cancel() {
+  if [[ ! -f "${WCE_FILE}" ]]; then
+      echo "Error: session not started"
+      exit 1
+  fi
+  if [[ -f "${WCE_SUSPEND_FILE}" ]]; then
+      echo "Error: session suspended"
+      exit 1
+  fi
+  echo "Canceling session: ${TODAY}"
   rm "${WCE_FILE}"
   export PS1="${PS1_ORIGINAL}"
 }
@@ -117,6 +149,9 @@ function main() {
       finish)
           finish
           ;;
+      cancel)
+        cancel
+        ;;
       suspend)
           suspend
           ;;
